# Attendance-system/backend/swagger.yaml
openapi: 3.0.3
info:
  title: DIAMS API
  description: |
    **Digital Intelligent Attendance Management System**

    REST API for managing attendance sessions, courses, and users.
    Supports three attendance methods: BLE beacon, QR code, and Manual,
    with automatic HMAC fallback when external microservices are unavailable.

    ## Authentication
    All protected routes require:
    ```
    Authorization: Bearer <JWT>
    ```
    Obtain a token via `POST /login`.

    ## Roles
    | Role | Description |
    |------|-------------|
    | `admin` | Full system access |
    | `prof` | Own courses only |
    | `ta` | Same as prof on assigned courses |
    | `student` | Mobile app only — limited to own data |

    ## Timezone
    All schedule times (`startTime`, `endTime`, `scheduledDay`) are **IST strings**.
    Lecture `scheduledTime` is stored and returned as **UTC**.
  version: "2.0.0"
  contact:
    name: DIAMS Backend
servers:
  - url: http://localhost:4040
    description: Local development
  - url: https://your-production-host
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Schedule:
      type: object
      properties:
        scheduledDay: { type: string, example: Monday }
        startTime:    { type: string, example: "09:00" }
        endTime:      { type: string, example: "09:50" }
        method:       { type: string, enum: [BLE, QRCode, Manual] }
        switch:       { type: boolean, description: Whether the method can be switched mid-session }
    Lecture:
      type: object
      properties:
        lectureUID:    { type: string }
        scheduledTime: { type: string, format: date-time, description: UTC }
        cancelled:     { type: boolean }
    Course:
      type: object
      properties:
        _id:         { type: string, example: CS101 }
        id:          { type: string, description: Alias for _id returned by API }
        name:        { type: string }
        department:  { type: string }
        slot:        { type: string, example: A }
        venue:       { type: string }
        startDate:   { type: string, format: date }
        endDate:     { type: string, format: date }
        instructors: { type: array, items: { type: string } }
        tas:         { type: array, items: { type: string } }
        lectures:    { type: array, items: { $ref: '#/components/schemas/Lecture' } }
        schedules:   { type: array, items: { $ref: '#/components/schemas/Schedule' } }
    CourseCreate:
      type: object
      required: [_id, name]
      properties:
        _id:         { type: string }
        name:        { type: string }
        department:  { type: string }
        slot:        { type: string }
        venue:       { type: string }
        startDate:   { type: string, format: date }
        endDate:     { type: string, format: date }
        instructors: { type: array, items: { type: string } }
        tas:         { type: array, items: { type: string } }

    # ── Session ───────────────────────────────────────────────────────────────
...
    Error:
      type: object
      properties:
        error: { type: string }
security:
  - bearerAuth: []

# ── Paths ───────────────────────────────────────────────────────────────────────
paths:
...
  /courses:
    get:
      tags: [Courses]
      summary: List courses for authenticated user
      parameters:
        - in: query
          name: user
          schema: { type: string }
          description: User ID — admin can omit to get all
      responses:
        "200":
          description: Array of courses
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Course' }

    post:
      tags: [Courses]
      summary: Create course (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CourseCreate' }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Course' }
  /courses/{courseId}:
    parameters:
      - in: path
        name: courseId
        required: true
        schema: { type: string }

    get:
      tags: [Courses]
      summary: Get course detail
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Course' }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    put:
      tags: [Courses]
      summary: Update course (admin)
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CourseCreate' }
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Course' }

    delete:
      tags: [Courses]
      summary: Delete course (admin)
      responses:
        "200":
          description: Deleted
  /courses/{courseId}/schedules:
    parameters:
      - in: path
        name: courseId
        required: true
        schema: { type: string }
    get:
      tags: [Courses]
      summary: Get schedules for course
      responses:
        "200":
          description: Schedule array
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Schedule' }
  /courses/{courseId}/schedule:
    parameters:
      - in: path
        name: courseId
        required: true
        schema: { type: string }

    post:
      tags: [Courses]
      summary: Add schedule entry (instructor only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Schedule' }
      responses:
        "201":
          description: Schedule added

    put:
      tags: [Courses]
      summary: Replace entire schedules array (instructor only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                schedules:
                  type: array
                  items: { $ref: '#/components/schemas/Schedule' }
      responses:
        "200":
          description: Schedules replaced
  /courses/{courseId}/schedule/{index}:
    parameters:
      - in: path
        name: courseId
        required: true
        schema: { type: string }
      - in: path
        name: index
        required: true
        schema: { type: integer }

    patch:
      tags: [Courses]
      summary: Update specific schedule entry (instructor only)
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Schedule' }
      responses:
        "200":
          description: Updated

    delete:
      tags: [Courses]
      summary: Remove schedule entry (instructor only)
      responses:
        "200":
          description: Deleted
  /courses/{courseId}/lectures/{lectureUID}/cancel:
    parameters:
      - in: path
        name: courseId
        required: true
        schema: { type: string }
      - in: path
        name: lectureUID
        required: true
        schema: { type: string }
    patch:
      tags: [Courses]
      summary: Cancel lecture — ends any active sessions immediately
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:        { type: string }
                  sessionsClosed: { type: integer }
        "404":
          description: Lecture not found
  /courses/{courseId}/lectures/{lectureUID}/uncancel:
    parameters:
      - in: path
        name: courseId
        required: true
        schema: { type: string }
      - in: path
        name: lectureUID
        required: true
        schema: { type: string }
    patch:
      tags: [Courses]
      summary: Reinstate cancelled lecture
      responses:
        "200":
          description: Reinstated

  # ── Sessions ────────────────────────────────────────────────────────────────
...